This should fix the "can't kill window" bugs. Well at least those I was able to reproduce. I've fixed the reference counting on borders, reference counts are now only modified where the references actually change.
PLEASE: if we don't want to end up dead in the water with the reference
counts, it'll be absolutely necessary to not just increment them here
and there. When you store a pointer in a data structure somewhere,
increment the counter. When you remove the pointer from the structure,
decrement. Do NOT modify the counters anywhere else.

Unfortunately I cannot enforce this in C.