Got rid of compiler warnings.
Added window spacing, so file icons will not appear underneath iconbars.
Currently, icons are placed twice as far from the left/top as the spacing.window.l,t designate. To compensate, the spacing is set to half the actual value. This needs fixing.

It seems as though there are a few other bugs in e_icon_apply_xy() also. (multiple line views don't seem to work quite right, icons on the bottom line are in the middle, and other lines seem right justified. a bit odd).


git-svn-id: http://svn.enlightenment.org/svn/e/trunk/e17/apps/e@5815 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33