- add more signals to icons / bg (need to properly rename them to make more sense) - add dynamic glob matches to typebuffer (can get slow with 10,000 files, as expected, should be faster when we optimize e_icon_layout) - add ability to start efm from e_remote (thanks ngc) - add some edje effects. - add a test idea for dynamic selection: when you bring up the typebuffer, everything goes into "dark mode", and your searches light up dynamically as they are found while you type.
- question: should A be equal to *A* in the typebuffer? ngc and rephorm
suggested this could be an option and if we do add a * manually then this
case will be invalidated.

- note: when browsing 10,000+ files, the slowness of the dir loading occurs
due to the fact that ecore_file's monitor will loop thru all the files and do
an insertion sort. raster said we can fix this with some work on the monitor
code.