Add a close button callback hook at the request of devilhorns.
If the callback returns zero, the dialog is NOT closed.

Seems that every user of config_dialog allocates E_Config_Dialog_View on
the stack and doesn't clear it, so I can't rely on the close_cfdata member
being NULL.  I currently set it to NULL in e_config_dialog_new() and if
you want to use it, set it in create_widgets();  I suspect that allocating
a structure on the stack that lives beyond the function call is just
asking for trouble.