add filemanager operation registry.
This huge commit will add e_fm2_op_registry, a central place that will
store all on-going operations. You will find that it will report when
operations are added, remove and when they change, for example when
they change progress.

Please notice that the recommended way to present information is to
add listener to specific entries, as opposed to use
E_EVENT_FM_OP_REGISTRY_CHANGED since it will be called immediately and
you do not have to filter which entry is being changed.

Entries will be associated with the e_fm object that originated
it. With that one can get the Ecore_X_Window and request the window to
be raised from somewhere else (ie: gadman/gadget). If object is
deleted (ie: window is closed), the pointer will be made NULL and it
will be a "windowless operation".

TO DO (I need someone to do those):

  - e_fwin: windows should present on-going operations as an overlay,
    Dave Andreolli already started such work, it's just a matter of
    using the new infrastructure.

  - create a new gadget that present all on-going operations and maybe
    a history. It would be nice to show the e_fwin of if the entry is
    clicked, as well as a cancel button and a progress bar. See
    "places" module, make them similar.




git-svn-id: http://svn.enlightenment.org/svn/e/trunk/e@39541 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33