turn off composite on segfault.
one of the most annoying problems while using bling is that when E
segfaults it will launch an alert window, but since composite manager
is using the same process/connection as e17, it will not die and e
will not compose as required, so you basically end with a black
screen.

this patch will forcely turn off manual redirection and you'll get the
white-screen-of-death, being able to restart e17.

future: if we really do composite in e18, then please handle this in a
better way.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk/e@39741 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33