fix annoying bug with config dialog window layout.
as spotted by manio, toggling between "Basic" and "Advanced" mode was
making the window move (just open Theme dialog and toggle).

Manio suggested a fix by keeping window center between both modes, but
I think that given our top-left to bottom-right reading (LTR) we
better keep the title bar (x,y) intact and avoid full rescan. This can
be weird in small screens where the dialog will not be centered
anymore and that will show easily, but on regular screens this should
not be an issue.

Also spotted that it was wrong when trying to avoid window overflow
(using x instead of y).

PS: I wonder why we destroy and create the window all over again! This
is really annoying, title bar will show the focus animation and you
get lost (need to rescan it again).




git-svn-id: http://svn.enlightenment.org/svn/e/trunk/e@41004 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33