fix bug with dialog resize.
This nasty bug was more evident with dialog reuse on configuration
Basic/Advanced toggle, on some resolutions like 1280x1024 it was
always reprodutible.

The problem spotted by Manio is that the dialog contents is resized
but the window is not. The bug follow is:

 * window is at some given size (640x512)
 * user toggles contents, calls e_dialog_show()
 * bg_object (window contents) is resized to MIN (!!!)
 * e_win_resize is called with MIN (so far, so good)
 * something calls e_util_win_auto_resize_fill(), size is back to 640x512
 * _e_border_eval() resizes to the last value 640x512, it's unchanged
   so _ecore_evas_x_resize() checks for size being the same and ignore,
   not even callking back fn_resize (user callback, _e_dialog_cb_resize).
 * window is at 640x512, while bg_object is at MIN, BUG!

One solution is to never change bg_object from outside
_e_dialog_cb_resize, like it's after this patch. The other is to
always call fn_resize even if it did not changed, which could lead to
lots of useless work.

By: manio
Confirmed: barbieri


git-svn-id: http://svn.enlightenment.org/svn/e/trunk/e@41515 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33